Added back
I added back the important files we needed. ( This will be yet another PR in the future since the new site has different files now ) I have no idea when 2.009 will come out so this will due for now but make it easier to just add to this when need be.